+++ Source/Iridio.Tests/CorrelateTests.cs
@@ -1,12 +1,13 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 using FluentAssertions;
+using Iridio.Tests.Tokenization;
 using Iridio.Tokenization;
 using Xunit;
 
 namespace Iridio.Tests
 {
-    public class UnitTest1
+    public class CorrelateTests
     {
         [Theory]
         [InlineData(new[] { 1 }, new[] { 1 })]
@@ -28,26 +29,6 @@ namespace Iridio.Tests
         {
             var actual = input.Correlate((a, b) => !TokenStream.InvalidTokenAdjacencyList.Contains((a, b)));
             actual.Should().BeEquivalentTo(expected);
-        }
-    }
-
-
-    public class TokenAdjacencyTests
-    {
-        [Theory]
-        [InlineData(SimpleToken.Text, SimpleToken.Text, false)]
-        [InlineData(SimpleToken.Integer, SimpleToken.Text, false)]
-        [InlineData(SimpleToken.Text, SimpleToken.Integer, false)]
-        [InlineData(SimpleToken.Double, SimpleToken.Integer, false)]
-        [InlineData(SimpleToken.Integer, SimpleToken.Double, false)]
-        [InlineData(SimpleToken.Integer, SimpleToken.Integer, false)]
-        [InlineData(SimpleToken.CloseBrace, SimpleToken.OpenBrace, true)]
-        public void Token_adjacency_is_valid(SimpleToken a, SimpleToken b, bool expected)
-        {
-            var invalidAdjacency = TokenStream.InvalidTokenAdjacencyList;
-
-            var isValid = !invalidAdjacency.Contains((a, b));
-            isValid.Should().Be(expected);
         }
     }
 }
